backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.3
httpx==0.27.2
python-dateutil==2.9.0
psycopg2-binary==2.9.9
bcrypt==4.2.0
gunicorn==22.0.0
